requirements.txt
faiss-cpu 
groq
langchain
langchain-groq
PyPDF2
langchain_google_genai
langchain
streamlit
langchain_community
python_dotenv